cmd/list: Add filter option, can search for bugs now
A new config option 'default_filter' is added. To get a list of bugs
assigned to the default user ('default_user' config option) and the
status of each bug is one of open or assigned, issue

    bugalert list --filter="status:OPEN|ASSIGNED"

Note the 'pipe' symbol, to denote 'OR'. Multiple filters can be passed separated
by a semicolon.

    bugalert list --filter="status:OPEN|ASSIGNED;priority:P1"

Signed-off-by: Santosh Sivaraj <santosh@fossix.org>